Created A2A namespace and moved appropriate items
