+++ SafeguardDotNet/SafeguardDotNet.csproj
@@ -24,4 +24,8 @@
     <PackageReference Include="Serilog" Version="2.7.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Event\" />
+  </ItemGroup>
+
 </Project>
